requirements.txt
load_dotenv
pytest
pymongo
mysql-connector-python
pathlib